+++ packages/table-render/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### 0.7.5
 
-- [+] `searchApi`为`Array`类型时, params 透出 tab
+- [+] `api`为`Array`类型时, params 透出 tab
 
 ### 0.7.0